requirements.txt
# FastAPI Core
fastapi==0.111.0
uvicorn==0.32.1
python-dotenv==1.0.1

# Database - Motor (async MongoDB driver)
motor==3.6.0
pymongo==4.9.1 

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Serialization
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.2.0

# LangChain & Groq (compatible with FastAPI + Python 3.11)
langchain==0.3.9
langchain-core==0.3.29
langchain-community==0.3.8
langchain-groq==0.2.0
# HTTP & Requests
httpx==0.28.0
requests==2.32.3

# Azure Storage
azure-storage-blob==12.24.0
azure-core==1.32.0

# Utilities
python-multipart==0.0.18

# OCR & Image Processing
easyocr==1.7.1
pdf2image==1.17.0
Pillow==11.0.0
numpy==1.26.4